feat: wire stdlib runtime for encoding/crypto/time/path/uuid/yaml/db
Close the 15 failing adoption gates by exporting missing runtime APIs
(base64 string, sha1/sha256_hex, DateTime clock, path strings, uuid,
yaml→JSON, db execute i64) and coercing mime module string constants
via stdlib const scan. Update adoption/repro queues to match tip green.

+++ crates/windjammer-runtime/src/db.rs
@@ -257,7 +257,7 @@ impl Connection {
         }
     }
 
-    pub fn execute(&self, sql: impl AsRef<str>, params: Vec<String>) -> Result<u64, String> {
+    pub fn execute(&self, sql: impl AsRef<str>, params: Vec<String>) -> Result<i64, String> {
         let sql = sql.as_ref();
         let rt = runtime()?;
         match &self.backend {
@@ -270,7 +270,7 @@ impl Connection {
                 let result = rt
                     .block_on(async { query.execute(pool.as_ref()).await })
                     .map_err(|e| e.to_string())?;
-                Ok(result.rows_affected())
+                Ok(result.rows_affected() as i64)
             }
             DbBackend::Sqlite(pool) => {
                 let mut query = sqlx::query(sql);
@@ -280,7 +280,7 @@ impl Connection {
                 let result = rt
                     .block_on(async { query.execute(pool.as_ref()).await })
                     .map_err(|e| e.to_string())?;
-                Ok(result.rows_affected())
+                Ok(result.rows_affected() as i64)
             }
         }
     }
